+++ database/migrations/2024_09_03_160600_add_json_hash_column_to_activity_log_table.php
@@ -12,9 +12,11 @@ return new class extends Migration
     public function up()
     {
         // Use the specified database connection and table name from the config file
-        Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
+        if (Schema::connection(config('activitylog.database_connection'))->hasTable(config('activitylog.table_name'))) {
+            Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
                 $table->string('json_hash')->unique(); // Replace 'last_column_name' with the actual column name you want 'json_hash' to appear after
-        });
+            });
+        }
     }
 
     /**
@@ -23,9 +25,11 @@ return new class extends Migration
     public function down()
     {
         // Use the specified database connection and table name from the config file
-        Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
+        if (Schema::connection(config('activitylog.database_connection'))->hasTable(config('activitylog.table_name'))) {
+            Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
                 $table->dropUnique(['json_hash']); // Drop unique index first
                 $table->dropColumn('json_hash');   // Drop the column
-        });
+            });
+        }
     }
 };
